♻️(recrutement-presentation) isolate RecrutementEtapeView and InitRecrutementEtapeView to recrutement_params.py for more clarity
